% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldapqp_grp.R
\name{check_AD_grp}
\alias{check_AD_grp}
\title{Checks whether a given user is member of an AD specific group}
\usage{
check_AD_grp(ldserver, bind_dn, bind_pass, base, user, group)
}
\arguments{
\item{bind_dn}{Bind DN used for querying}

\item{bind_pass}{Bind password}

\item{base}{Base DN as entrypoint in the three}

\item{user}{User whose membership to look for}

\item{group}{The AD group to check for}

\item{server}{The AD server to query}
}
\value{
Boolean
}
\description{
Function to verify whether a given user is member of a specific AD group
}
\examples{
check_AD_grp('ldap://myldapserver.com:389',
            'cn=bind_user,dc=example,dc=com',
            'bind_password',
            'cn=base,dc=example,dc=com',
            'user','group')
}
