% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldapqp_grp.R
\name{check_grp}
\alias{check_grp}
\title{Checks whether a given user is member of a specific group}
\arguments{
\item{server}{The ldap server to query}

\item{bind_dn}{Bind DN used for querying}

\item{bind_pass}{Bind password}

\item{base}{Base DN as entrypoint in the three}

\item{user}{User whose membership to look for}

\item{group}{The group to check for}
}
\value{
Boolean
}
\description{
Function to verify whether a given user is member of a specific group
}
\examples{
check_grp('ldap://myldapserver.com:389',
            'cn=bind_user,dc=example,dc=com',
            'bind_password',
            'cn=base,dc=example,dc=com',
            'user','group')
}
